refactor: remove logic duplication for redirect, add QueryParamshandler.tsx
