--- src/pages/auth/signin/ui/AuthRedirectHandler.tsx
@@ -1,7 +1,0 @@
-'use client';
-import { useAuthRedirectMessage } from '../model/useAuthRedirectMessage';
-
-export function AuthRedirectHandler() {
-  useAuthRedirectMessage();
-  return null;
-}
