+++ src/shared/ui/index.ts
@@ -40,3 +40,4 @@ export * from './Empty';
 export * from './ScrollArea';
 export * from './Kanban';
 export * from './icon';
+export { QueryParamsHandler } from './QueryParamsHandler';
